feat(team): add endpoint to retrieve team members of the authenticated user's project

+++ src/main/java/com/hr/workwave/repo/LeaveRequestRepository.java
@@ -59,7 +59,7 @@ public interface LeaveRequestRepository extends JpaRepository<LeaveRequest, Long
 
 
     @Query("SELECT COUNT(lr) FROM LeaveRequest lr WHERE lr.leave_type = 'HOME_OFFICE' AND lr.start_date <= :date AND lr.end_date >= :date AND lr.user.project.id = :projectId")
-    int countHomeOfficeRequestsOnDateAndProject(@Param("date") LocalDateTime date, @Param("projectId") BigInteger projectId);
+    int countHomeOfficeRequestsOnDateAndProject(@Param("date") LocalDateTime date, @Param("projectId") Long projectId);
 
     @Query("SELECT CASE WHEN COUNT(lr) > 0 THEN true ELSE false END FROM LeaveRequest lr " +
             "WHERE lr.user.id = :userId AND lr.start_date <= :weekEnd AND lr.end_date >= :weekStart " +
